Final UML Update Before Submition
This is the last update to the UML Diagram before submiting the design doc.
The addToLayPhase method in Phase state was modified to have an extra param
and the drawAction class was modified to have a boolean variable saying if
the draw action was from the deck or not. This will allow drawing from either
the deck or the discard pile.