Commit the Phase10 Starter Project
This commit is commiting the starting point for our Phase10 game. I took the
SJProj, and then went in and refactored all of the names for the packages and
files so that they all now say phase10 and all refereneces have been updated.
I left all of the files the same and it will play the SJ game so we need to
go in and basically delete all of those files under the phase10 package but
I left them there for now incase people wanted to get an idea of how Vegdahl
structured his game to start with. There may be some things that can be
reused, but soon I will just delete all files specific to the game and we'll
start over from just the framework.