Start of the Game
I have started to work on the basics that will be needed for the game. I have
created a Hand class that will be used to hold the cards in a users hand.
This class is not done, as it will probably inherit from Deck and deck will
be modified to contain more methods or something of the like. At this point I
am unsure. I have also created a PhaseState class. It currently has nothing
in it, as I needed the support classes before I could build much. I have also
edited all of the files in the edu/up.cs301.card package so that now instead
of suites, there are colors and the ranks have been modified to have the
correct ranks for the cards, including skip and wild. The use of these will
be looked into later.