Working Starting Point
I have structured and built most of the files that we need for making the
game. Most of the main files are here and will compile and allow you to run
the game on the tablet. They are mostly dumbie classes for not now with not
to much in them. It will all compile though and you should be able to get
past the main starting point of the app. This allows for tests to be built
and to run on the code that you are making. A test package has also been made
so that everyone can test their classes as they go. It should be all set up
correctly and pointing at the correct things. Model your test classes after
the ond that's already in there.
Good luck guys, and lets hope it goes smoothly. Let me know if you have any
questions.