Bug in Phase State Found and Fixed
There was a bug in Phase State when the initial set up of the game was run.
The game was crashing on start up of the actual game. This has been fixed and
the changes will be pushed to MaxTested and then immediately merged to master
and then pushed.
Please pull and merge from master into your branch as soon as possible as the
changes will allow you to test and run the program.