Fix for WildCard hitting, as well as fixes for after a round is over
Changes made to allow wild card hitting. There are also some fixes for the
the game not progressing correctly. It now looks to be progressing correctly
in that each player is on the correct Phase, the scores are added properly
and the board and everything is reset. The Gui, however does not acurately
show this at the moment. Will work on this next.

Sorry that the changes were from me in dan's branch.