A Few Bug Fixes
Fixed bugs with network play. Now have a network player working. There were
a few bugs that occured after network play started working. There was a crash
on playing a skip card, and a bug with hitting on players phases that caused
both of the player's phases to be the same. This issue would also have
occured on the local game only but we could never get a computer to phase.
These are both fixed now. The only issues left for network play I believe are
that the laid Phases aren't being updated correctly... If a phase is hit on
the remote player does not see this. I have no idea why. I will look into
this.